legger til checbox for valg av kompakt visning. Flytter tabeller til egen fil
