+++ src/frontend/App/typer/infotrygd.ts
@@ -1,5 +1,4 @@
 import { Stønadstype } from './behandlingstema';
-import { InfotrygdPeriodeMedFlereEndringer } from '../../Komponenter/Infotrygd/typer';
 
 export interface InfotrygdPerioderResponse {
     overgangsstønad: Perioder;
@@ -376,197 +375,3 @@ export enum InfotrygdSakUndervalg {
     OL = 'OL',
     OR = 'OR',
 }
-
-export const infotrygdperioder: InfotrygdPeriodeMedFlereEndringer[] = [
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.ANNULERT,
-        sakstype: Sakstype.MASKINELL_SATSOMREGNING,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.BARN_UNDER_1_3_ÅR,
-        aktivitetstype: InfotrygdAktivitetstype.IKKE_I_AKTIVITET,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.ENDRING_BEREGNINGSGRUNNLAG,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.YRKESRETTET_AKTIVITET_BARN_FYLT_1_3_ÅR,
-        aktivitetstype: InfotrygdAktivitetstype.I_ARBEID,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.FØRSTEGANGSVEDTAK,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.UNNTAK_FRA_KRAV_TIL_YRKESRETTET_AKTIVITET,
-        aktivitetstype: InfotrygdAktivitetstype.BRUKERKONTAKT,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.G_REGULERING,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.UTVIDELSE_NØDVENDIG_UTDANNING,
-        aktivitetstype: InfotrygdAktivitetstype.TILMELDT_SOM_REELL_ARBEIDSSØKER,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.NY,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.PÅVENTE_SKOLESTART_ARBEID_TILSYNSPLASS,
-        aktivitetstype: InfotrygdAktivitetstype.TILMELDT_SOM_REELL_ARBEIDSSØKER,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.OPPHØRT,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.YRKESRETTET_AKTIVITET,
-        aktivitetstype: InfotrygdAktivitetstype.TILMELDT_SOM_REELL_ARBEIDSSØKER,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.SATSENDRING,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.FORBIGÅENDE_SYKDOM,
-        aktivitetstype: InfotrygdAktivitetstype.TILMELDT_SOM_REELL_ARBEIDSSØKER,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.UAKTUELL,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.SÆRLIG_TILSSYNSKREVENDE_BARN,
-        aktivitetstype: InfotrygdAktivitetstype.TILMELDT_SOM_REELL_ARBEIDSSØKER,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.OVERTFØRT_NY_LØSNING,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad: InfotrygdOvergangsstønadKode.ETABLERER_EGEN_VIRKSOMHET,
-        aktivitetstype: InfotrygdAktivitetstype.TILMELDT_SOM_REELL_ARBEIDSSØKER,
-        brukerId: 'S12343243',
-    },
-    {
-        initiellKode: Kode.NY,
-        stønadId: 2,
-        vedtakId: 2,
-        vedtakstidspunkt: '2022-05-25T16:14:47.883255',
-        stønadFom: '2022-05-25T16:14:47.883255',
-        stønadTom: '2022-05-25T16:14:47.883255',
-        opphørsdato: '2022-05-25T16:14:47.883255',
-        månedsbeløp: 2000,
-        engangsbeløp: 10000,
-        inntektsgrunnlag: 200000,
-        samordningsfradrag: 200,
-        utgifterBarnetilsyn: 2000,
-        kode: Kode.NY,
-        sakstype: Sakstype.SØKNAD,
-        kodeOvergangsstønad:
-            InfotrygdOvergangsstønadKode.FORTSATT_INNVILGET_TROSS_VARSEL_OM_OPPHØR_PGA_SAMBOER,
-        aktivitetstype: InfotrygdAktivitetstype.TILMELDT_SOM_REELL_ARBEIDSSØKER,
-        brukerId: 'S12343243',
-    },
-];
